% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expand.R
\name{studydata}
\alias{studydata}
\title{Experience Study Data Processing}
\usage{
studydata(data, study.end, policy.start, death.date, sex, issue.age, keyp)
}
\arguments{
\item{data}{data.table structure}

\item{study.end}{study end date}

\item{policy.start}{policy start date}

\item{death.date}{death date}

\item{sex}{gender}

\item{issue.age}{age at policy issue}

\item{keyp}{unique ID}
}
\value{
data expanded by duration w/exposures & expecteds using VBT 2015 and 15% smoker assumption
}
\description{
Experience Study Data Processing
}
\examples{
dt <- studydata(dt, "2019-12-17", "Start", "Death", "Gender", "Age", "Id")
}
\author{
Conrad Chen
}
