% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data for study processing}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 100 rows and 5 columns.
}
\usage{
data(test_data)
}
\description{
Test data for study processing
}
\keyword{datasets}
